[LLT-4202] Make lookup return error in case of failure on forward
Make the forward error visible instead of just filling the response
buffer with an empty response.

The inieal problem was that if the forward lookup from trust-dns fails,
the failure is not indicated to libtelio. Instead the response buffer
was filled with an empty lookup response. This, combined with the
doubled retry mechanis (OS and trust-dns) produced some issues on IOS.

This commit also removes the compatibility checks from the workflows. It
started to fail after the version update. I created LLT-4544 to fix it.